refactor: import the upload orchestration from services/apiMedia and drop the fileUpload re-export shim (#3452)
The four network-calling upload helpers moved to services/apiMedia.js
behind a backward-compatible re-export from utils/fileUpload.js so no
caller had to change mid-audit. Repoints the four real call sites at
apiMedia and removes the shim, so utils/fileUpload.js is once again pure
helpers and constants with no reach into the network layer.
